Adds DisableWithable utility module & traits for shared access to disable_with option and trait for button and dropdown item (and others as we come across it).
