+++ app/views/nfg_ui/patterns/dropdowns/_dropdown.html.haml
@@ -1,4 +1,4 @@
 %div{ dropdown.html_options }
   - if dropdown.button.present?
-    = ui.nfg :dropdown_toggle, body: dropdown.button, theme: dropdown.theme, as: :button, offset: dropdown.offset, icon: dropdown.icon
+    = ui.nfg :dropdown_toggle, body: dropdown.button, theme: dropdown.theme, as: :button, offset: dropdown.offset
   = dropdown.body
